requirements.txt
flask>=2.3
requests>=2.31
beautifulsoup4>=4.12
python-telegram-bot>=20.3